lib/util: major enhancements:
msg_decoder: support SRC filtering

trace_decoder:
  - fix incorrect handling of HREPEAT
  - sign-extend X-ADDR to hwcfg.addr_bits
  - add NEXUSRV_Trace_Event_(In)DirectSync events

util:
  - fix insn decoding (capstone gives j/jal/jr/jalr)
  - ignore unsupported capstone disasm and continue
  - proper handling of linux kcore/ucore and symbols

+++ util/misc.c
@@ -59,7 +59,7 @@ int open_seek_file(char *filename, int oflags) {
         fd = open(filename, oflags);
     if (fd < 0)
         error(-1, errno, "Failed to open file %s", filename);
-    if (seek_file(fd, file_offset) != file_offset)
+    if (seek_file(fd, file_offset) != (ssize_t)file_offset)
         error(-1, errno, "Failed to seek file %s", filename);
     return fd;
 }
